requirements.txt
python-telegram-bot==20.0
sqlalchemy==2.0.0
alembic==1.11.0
python-dotenv==0.21.0